backend/requirements.txt
# SlideForge Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6  # Required for file uploads

# Supabase (PostgreSQL + Storage + Auth)
supabase==2.5.0
httpx==0.26.0

# Google AI
google-generativeai==0.3.1

# Data Models
pydantic>=2.10.0
pydantic-settings>=2.7.0

# JWT Authentication
PyJWT==2.8.0
cryptography==41.0.7

# HTTP Client (for external APIs)
aiohttp>=3.10.0

# Date/Time
python-dateutil==2.8.2

# Social Media APIs
# Ayrshare (when ready)
# requests==2.31.0

# Image Generation (optional)
# pillow==10.1.0
# requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
stripe==5.0.0
